Fix clang -Wshorten-64-to-32 warnings in port/ subdir (https://trac.osgeo.org/gdal/ticket/6218)
git-svn-id: https://svn.osgeo.org/gdal/trunk@31482 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_error.cpp
@@ -268,7 +268,7 @@ void    CPLErrorV( CPLErr eErrClass, CPLErrorNum err_no, const char *fmt,
         if ( psCtx->psHandlerStack != NULL &&
              EQUAL(CPLGetConfigOption( "CPL_ACCUM_ERROR_MSG", "" ), "ON"))
         {
-            nPreviousSize = strlen(psCtx->szLastErrMsg);
+            nPreviousSize = static_cast<int>(strlen(psCtx->szLastErrMsg));
             if (nPreviousSize)
             {
                 if (nPreviousSize + 1 + 1 >= psCtx->nLastErrMsgMax)
